+++ picopt/handlers/image.py
@@ -115,9 +115,6 @@ class ImageHandler(Handler, metaclass=ABCMeta):
         self, _exec_args: tuple[str, ...], old_path: Path, new_path: Path
     ) -> Path:
         """Internally convert unhandled formats to uncompressed png for cwebp."""
-        # It's faster to create a undercompressed png than anything else
-        if self.input_file_format.format_str == PngImageFile.format:
-            return old_path
         return self.pil2native(
             self.EMPTY_EXEC_ARGS,
             old_path,